pyproject.toml
[project]
name = "gemini-api-file-search-tool"
version = "0.1.0"
description = "Streamlit PDF chat using Google Gemini File Search"
readme = "README.md"
authors = [{ name = "Gemini PDF Chat Maintainers" }]
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.38.0",
  "google-genai>=0.2.2",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
]

[tool.uv]
package = true

[tool.uv.sources]
# defaults to PyPI; add private indexes here if needed
